replace makeStyle with emotion wip
