+++ generators/app/templates/infrastructure/src/assets/jss/components/appStyle.js
@@ -1,77 +1,38 @@
+import { styled } from '@mui/material'
 import styles from 'assets/jss/styles'
+import { includes } from 'ramda'
 
-const appStyle = theme => {
-  const { drawerWidth, rightDrawerWidth, drawerMiniWidth, defaultFont /*,transition*/ } = styles(theme)
-
-  return {
-    wrapper: {
-      ...defaultFont,
-      backgroundColor: theme.palette.background.main,
-      color: theme.palette.primary.main,
-      position: 'relative',
-      top: '0',
-      height: '100vh',
-      '&:after': {
-        display: 'table',
-        clear: 'both',
-        content: '" "'
-      }
-    },
-    mainPanel: {
-      [theme.breakpoints.up('md')]: {
-        width: `calc(100% - ${drawerWidth}px)`
-      },
-      overflow: 'auto',
-      overflowX: 'hidden',
-      position: 'relative',
-      float: 'right',
-      maxHeight: '100%',
-      height: '100%',
-      width: '100%',
-      overflowScrolling: 'touch'
-    },
-    mainPanelSidebarMini: {
-      [theme.breakpoints.up('md')]: {
-        width: `calc(100% - ${drawerMiniWidth}px)`
-      }
-    },
-    appBarShift: {
-      [theme.breakpoints.up('md')]: {
-        width: `calc(100% - ${drawerWidth + rightDrawerWidth}px)`
-      }
-    },
-    appBarShiftMini: {
-      [theme.breakpoints.up('md')]: {
-        width: `calc(100% - ${drawerMiniWidth + rightDrawerWidth}px)`
-      }
-    },
-    'appBarShift-right': {
-      [theme.breakpoints.up('md')]: {
-        marginRight: rightDrawerWidth
-      }
-    },
-
-    mainPanelWithPerfectScrollbar: {
-      overflow: 'hidden !important'
-    },
-    progressBar: {
-      height: 2,
-      position: 'absolute',
-      zIndex: 2000,
-      top: 0,
-      left: 0,
-      width: '100%',
-      bottom: 0
-    },
-    stickySideButtons: {
-      position: 'sticky',
-      top: 0,
-      '@media (max-width: 480px)': {
-        position: 'inherit',
-        zIndex: 1030
-      }
-    }
+export const Container = styled('div')(({ theme }) => ({
+  ...styles(theme)?.defaultFont,
+  backgroundColor: theme.palette.background.main,
+  color: theme.palette.primary.main,
+  position: 'relative',
+  top: '0',
+  height: '100vh',
+  '&:after': {
+    display: 'table',
+    clear: 'both',
+    content: '" "'
   }
-}
+}))
 
-export default appStyle
+export const Content = styled('div', {
+  shouldForwardProp: prop => !includes(prop, ['drawerOpen'])
+})(({ theme, drawerOpen }) => ({
+  [theme.breakpoints.up('md')]: {
+    width: `calc(100% - ${styles(theme)?.drawerWidth}px)`
+  },
+  overflow: 'auto',
+  overflowX: 'hidden',
+  position: 'relative',
+  float: 'right',
+  maxHeight: '100%',
+  height: '100%',
+  width: '100%',
+  overflowScrolling: 'touch',
+  ...(!drawerOpen && {
+    [theme.breakpoints.up('md')]: {
+      width: `calc(100% - ${styles(theme)?.drawerMiniWidth}px)`
+    }
+  })
+}))
