--- generators/app/templates/infrastructure/src/assets/jss/components/mainStyle.js
@@ -1,13 +1,0 @@
-const mainStyle = (_theme, { fullWidth }) => {
-  return {
-    container: {
-      padding: fullWidth ? 0 : '20px 40px',
-      minHeight: 'calc(100vh - 123px)',
-      '@media (max-width: 480px)': {
-        padding: '15px'
-      }
-    }
-  }
-}
-
-export default mainStyle
