+++ generators/app/templates/infrastructure/src/components/layout/sidebar/Sidebar.js
@@ -1,85 +1,53 @@
 import React, { useRef } from 'react'
+import { withRouter } from 'react-router-dom'
+import { useTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
-
+import { Hidden } from '@mui/material'
 import SimpleBar from 'simplebar-react'
 import 'simplebar/dist/simplebar.min.css'
 
-import { Drawer, Hidden } from '@mui/material';
-import { makeStyles} from 'tss-react/mui'
-import { Typography } from '@totalsoft_oss/rocket-ui.core'
-import { useTranslation } from 'react-i18next'
 import { env } from 'utils/env'
-
+import UserMenu from 'components/menu/user/UserMenu'
 import Menu from 'components/menu/Menu'
-import UserMenu from '../menu/UserMenu'
-import sidebarStyle from 'assets/jss/components/sidebarStyle'
-import cx from 'classnames'
-import { withRouter } from 'react-router-dom'
 import { sidebarWrapperHeight } from 'utils/constants'
-
-const useStyles = makeStyles()(sidebarStyle)
+import { Drawer, SidebarRef, StyledImage, StyledLogo, StyledLogoDefault, StyledLogoMini, Typography } from './SidebarStyle'
 
 // We've created this component so we can have a ref to the wrapper of the links that appears in our sidebar.
 // This was necessary so that we could initialize PerfectScrollbar on the links.
 // There might be something with the Hidden component from material-ui, and we didn't have access to
 // the links, and couldn't initialize the plugin.
-function SidebarWrapper({ className, children }) {
+function SidebarWrapper({ children, drawerOpen }) {
   const sidebarWrapperRef = useRef()
 
   return (
-    <div className={className} ref={sidebarWrapperRef}>
+    <SidebarRef ref={sidebarWrapperRef} drawerOpen={drawerOpen}>
       <SimpleBar style={{ height: sidebarWrapperHeight, overflowX: 'hidden' }}>{children}</SimpleBar>
-    </div>
+    </SidebarRef>
   )
 }
 
 SidebarWrapper.propTypes = {
-  className: PropTypes.string.isRequired,
-  children: PropTypes.array.isRequired
+  children: PropTypes.array.isRequired,
+  drawerOpen: PropTypes.bool.isRequired
 }
 
 function Sidebar({ logo, logoText, drawerOpen, changeLanguage, closeDrawer, withGradient }) {
-  const { classes}  = useStyles()
   const { i18n, t } = useTranslation()
 
-  const logoNormal =
-    classes.logoNormal +
-    ' ' +
-    cx({
-      [classes.logoNormalSidebarMini]: !drawerOpen
-    })
-
-  var brand = (
-    <div className={classes.logo}>
-      <a href='/' className={classes.logoMini}>
-        <img src={logo} alt='logo' className={classes.img} />
-      </a>
+  const brand = (
+    <StyledLogo>
+      <StyledLogoMini href='/'>
+        <StyledImage src={logo} alt='logo' />
+      </StyledLogoMini>
       {logoText && (
-        <a href='/' className={logoNormal}>
+        <StyledLogoDefault href='/' drawerOpen={drawerOpen}>
           {logoText}
-        </a>
+        </StyledLogoDefault>
       )}
-    </div>
+    </StyledLogo>
   )
 
-  var appVersion = (
-    <Typography className={classes.appVersion} variant={'caption'}>
-      {`${t('BuildVersion')} ${env.REACT_APP_VERSION}`}
-    </Typography>
-  )
-
-  const drawerPaper =
-    classes.drawerPaper +
-    ' ' +
-    cx({
-      [classes.drawerPaperMini]: !drawerOpen
-    })
-  const sidebarWrapper =
-    classes.sidebarWrapper +
-    ' ' +
-    cx({
-      [classes.drawerPaperMini]: !drawerOpen
-    })
+  const appVersion = <Typography variant={'caption'}>{`${t('BuildVersion')} ${env.REACT_APP_VERSION}`}</Typography>
 
   return (
     <div>
@@ -88,16 +56,14 @@ function Sidebar({ logo, logoText, drawerOpen, changeLanguage, closeDrawer, with
           variant='temporary'
           anchor='right'
           open={drawerOpen}
-          classes={{
-            paper: drawerPaper + ' ' + classes.themeBackground
-          }}
           onClose={closeDrawer}
           ModalProps={{
             keepMounted: true // Better open performance on mobile.
           }}
+          drawerOpen={drawerOpen}
         >
           {brand}
-          <SidebarWrapper className={sidebarWrapper}>
+          <SidebarWrapper drawerOpen={drawerOpen}>
             <UserMenu drawerOpen={drawerOpen} changeLanguage={changeLanguage} language={i18n.language} withGradient={withGradient} />
             <Menu drawerOpen={drawerOpen} withGradient={withGradient} />
           </SidebarWrapper>
@@ -105,16 +71,9 @@ function Sidebar({ logo, logoText, drawerOpen, changeLanguage, closeDrawer, with
         </Drawer>
       </Hidden>
       <Hidden smDown>
-        <Drawer
-          anchor='left'
-          variant='permanent'
-          open={drawerOpen}
-          classes={{
-            paper: drawerPaper + ' ' + classes.themeBackground
-          }}
-        >
+        <Drawer anchor='left' variant='permanent' open={drawerOpen} drawerOpen={drawerOpen}>
           {brand}
-          <SidebarWrapper className={sidebarWrapper}>
+          <SidebarWrapper drawerOpen={drawerOpen}>
             <UserMenu drawerOpen={drawerOpen} changeLanguage={changeLanguage} language={i18n.language} withGradient={withGradient} />
             <Menu drawerOpen={drawerOpen} withGradient={withGradient} />
           </SidebarWrapper>
